Rename datascience to datascience_modules (#10525)
* Rename datascience to datascience_modules

* Change name to something even harder to get wrong
