+++ pythonFiles/tests/ipython/scripts.py
@@ -79,7 +79,9 @@ def get_data_frame_info(variables, name, capsys):
     varJson = find_variable_json(variables, name)
     path = os.path.dirname(os.path.abspath(__file__))
     file = os.path.abspath(
-        os.path.join(path, "../../datascience/getJupyterVariableDataFrameInfo.py")
+        os.path.join(
+            path, "../../vscode_datascience_helpers/getJupyterVariableDataFrameInfo.py"
+        )
     )
     keys = dict([("_VSCode_JupyterTestValue", json.dumps(varJson))])
     if execute_script(file, keys):
@@ -92,7 +94,9 @@ def get_data_frame_info(variables, name, capsys):
 def get_data_frame_rows(varJson, start, end, capsys):
     path = os.path.dirname(os.path.abspath(__file__))
     file = os.path.abspath(
-        os.path.join(path, "../../datascience/getJupyterVariableDataFrameRows.py")
+        os.path.join(
+            path, "../../vscode_datascience_helpers/getJupyterVariableDataFrameRows.py"
+        )
     )
     keys = dict(
         [
