+++ pythonFiles/vscode_datascience_helpers/getJupyterKernelspecVersion.py
