+++ pythonFiles/vscode_datascience_helpers/getJupyterVariableList.py
