+++ src/client/datascience/constants.ts
@@ -12,7 +12,7 @@ export const DefaultTheme = 'Default Light+';
 export const JUPYTER_OUTPUT_CHANNEL = 'JUPYTER_OUTPUT_CHANNEL';
 
 // Python Module to be used when instantiating the Python Daemon.
-export const PythonDaemonModule = 'datascience.jupyter_daemon';
+export const PythonDaemonModule = 'vscode_datascience_helpers.jupyter_daemon';
 
 // List of 'language' names that we know about. All should be lower case as that's how we compare.
 export const KnownNotebookLanguages: string[] = [
